refactor(product): 재고 복원 로직 이벤트 기반으로 변경
- 주문 서비스가 재고복원을 직접 하지 않도록 함
