+++ apps/commerce-api/src/main/java/com/loopers/application/order/OrderFacade.java
@@ -3,8 +3,8 @@ package com.loopers.application.order;
 import com.loopers.domain.coupon.CouponEntity;
 import com.loopers.domain.coupon.CouponService;
 import com.loopers.domain.order.OrderEntity;
-import com.loopers.domain.order.OrderPlacedEvent;
 import com.loopers.domain.order.OrderService;
+import com.loopers.domain.order.event.OrderPlacedEvent;
 import com.loopers.domain.payment.PaymentEntity;
 import com.loopers.domain.payment.PaymentService;
 import com.loopers.domain.product.ProductEntity;
