+++ apps/commerce-api/src/main/java/com/loopers/domain/order/OrderService.java
@@ -1,11 +1,12 @@
 package com.loopers.domain.order;
 
+import com.loopers.domain.order.event.OrderFailedEvent;
 import com.loopers.domain.payment.PaymentStatus;
-import com.loopers.domain.product.ProductEntity;
 import com.loopers.domain.product.ProductService;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -15,6 +16,7 @@ public class OrderService {
 
     private final OrderRepository orderRepository;
     private final ProductService productService;
+    private final ApplicationEventPublisher eventPublisher;
 
     public OrderEntity save(OrderEntity order) {
         return orderRepository.save(order);
@@ -30,39 +32,24 @@ public class OrderService {
             return;
         }
 
-        OrderEntity order = findById(orderId);
+        OrderEntity order = orderRepository.findByIdWithPessimisticLock(orderId)
+                .orElseThrow(() -> new EntityNotFoundException("Order not found: " + orderId));
+
         switch (status) {
             case SUCCESS -> order.complete();
             case FAILED -> {
                 order.fail();
-                restoreStock(orderId);
+                if (order.isStockDeducted()) {
+                    eventPublisher.publishEvent(new OrderFailedEvent(orderId));
+                }
             }
             case CANCELED -> {
                 order.cancel();
-                restoreStock(orderId);
+                if (order.isStockDeducted()) {
+                    eventPublisher.publishEvent(new OrderFailedEvent(orderId));
+                }
             }
         }
         save(order);
-    }
-
-    private void restoreStock(Long orderId) {
-        log.info("재고 복원 시작: orderId={}", orderId);
-
-        OrderEntity order = orderRepository.findByIdWithPessimisticLock(orderId)
-                .orElseThrow(() -> new EntityNotFoundException("주문을 찾을 수 없습니다."));
-
-        if (!order.isStockDeducted()) {
-            log.info("재고가 차감된 기록이 없어 복원을 건너뜁니다: orderId={}", orderId);
-            return;
-        }
-
-        order.getItems().forEach(item -> {
-            ProductEntity product = productService.findById(item.getProductId());
-            product.increaseStock(item.getQuantity());
-        });
-
-        order.markStockAsRestored();
-        orderRepository.save(order);
-        log.info("재고 복원 완료: orderId={}", orderId);
     }
 }
