+++ apps/commerce-api/src/main/java/com/loopers/domain/order/OrderStatus.java
@@ -7,7 +7,7 @@ import com.loopers.util.EnumUtils;
 public enum OrderStatus {
     PENDING,
     COMPLETED,
-    CANCELLED,
+    CANCELED,
     FAILED;
 
     @JsonValue
