+++ apps/commerce-api/src/main/java/com/loopers/domain/order/event/OrderFailedEvent.java
@@ -1,0 +1,6 @@
+package com.loopers.domain.order.event;
+
+public record OrderFailedEvent(
+        Long orderId
+) {
+}
